This commit was generated by cvs2svn to compensate for changes in r6815, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@6816 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/rsuser.c
@@ -102,11 +102,11 @@ static int set_up_enabled(resource_value_t v) {
 
     rsuser_enabled = newval;
     if(newval) {
- 	char_clk_ticks = 10.0 * cycles_per_sec / ((double)newval);
+ 	char_clk_ticks = (int)(10.0 * cycles_per_sec / ((double)newval));
     } else {
 	char_clk_ticks = RSUSER_TICKS;
     }
-    bit_clk_ticks = char_clk_ticks / 10.0;
+    bit_clk_ticks = (int)((double)(char_clk_ticks) / 10.0);
 #ifdef DEBUG
     log_debug("RS232: %d cycles per char (cycles_per_sec=%ld).",
               char_clk_ticks, cycles_per_sec);
